+++ src/commonTest/kotlin/io/github/z4kn4fein/semver/samples/ConstraintSamples.kt
@@ -16,7 +16,10 @@ class ConstraintSamples {
             "~1.0",
             "^1.x",
             "1.1.0 - 1.2.*",
-            ">=1.1.0 <3 || =0.1 || 5 - 6"
+            ">=1.1.0 <3 || =0.1 || 5 - 6",
+            "v1",
+            "v3 - v4",
+            ">=v2.3"
         )
 
         constraints.forEach { println("[$it]: [${it.toConstraint()}]") }