+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2022-01-23
+### Added
+- New `withoutSuffixes()` extension for `Version` which makes a copy without the pre-release and build metadata identities.
+- Support for parsing loose versions.
+- Constraints are now recognising the `v` prefix in conditions like: `>=v1.2`.
+
+### Fixed
+- During version incrementation the `preRelease` argument was not validated.
+
 ## [1.1.0] - 2022-01-16
 ### Added
 - New incrementation method `inc()`.